Create x86 emulator that uses kernel instead of kernel emulator
