+++ crates/pcode/src/arch/x86/emulator.rs
@@ -2,49 +2,63 @@ use std::rc::Rc;
 
 use libsla::{OpCode, PcodeInstruction, PseudoOp, Sleigh};
 
-use super::model::LinuxModel;
 use crate::emulator::{ControlFlow, Error, PcodeEmulator, Result, StandardPcodeEmulator};
+use crate::kernel::model::{KernelModel, NoKernel};
 use crate::mem::VarnodeDataStore;
 
 #[repr(u64)]
 enum CallOtherOps {
-    Segment = 0x00,
-    In = 0x01,
-    Out = 0x02,
-    SysEnter = 0x03,
-    SysExit = 0x04,
+    _Segment = 0x00,
+    _In = 0x01,
+    _Out = 0x02,
+    _SysEnter = 0x03,
+    _SysExit = 0x04,
     SysCall = 0x05,
     Lock = 0x11,
 }
 
 #[derive(Debug)]
-pub struct LinuxEmulator<S> {
+pub struct EmulatorX86<S, K = NoKernel>
+where
+    S: Sleigh,
+    K: KernelModel,
+{
     sleigh: Rc<S>,
-    linux: LinuxModel,
+    kernel: K,
     emulator: StandardPcodeEmulator,
 }
 
-impl<S> Clone for LinuxEmulator<S> {
+impl<S: Sleigh, K: KernelModel + Clone> Clone for EmulatorX86<S, K> {
     fn clone(&self) -> Self {
         Self {
             sleigh: self.sleigh.clone(),
-            linux: self.linux.clone(),
+            kernel: self.kernel.clone(),
             emulator: self.emulator.clone(),
         }
     }
 }
 
-impl<S: Sleigh> LinuxEmulator<S> {
+impl<S: Sleigh, K: KernelModel + Default> EmulatorX86<S, K> {
     pub fn new(sleigh: Rc<S>) -> Self {
         Self {
             emulator: StandardPcodeEmulator::new(sleigh.address_spaces()),
-            linux: Default::default(),
+            kernel: Default::default(),
             sleigh,
         }
     }
 }
 
-impl<S: Sleigh> PcodeEmulator for LinuxEmulator<S> {
+impl<S: Sleigh, K: KernelModel> EmulatorX86<S, K> {
+    pub fn with_kernel(sleigh: Rc<S>, kernel: K) -> Self {
+        Self {
+            emulator: StandardPcodeEmulator::new(sleigh.address_spaces()),
+            kernel,
+            sleigh,
+        }
+    }
+}
+
+impl<S: Sleigh, K: KernelModel> PcodeEmulator for EmulatorX86<S, K> {
     fn emulate<M: VarnodeDataStore>(
         &mut self,
         memory: &mut M,
@@ -67,10 +81,7 @@ impl<S: Sleigh> PcodeEmulator for LinuxEmulator<S> {
                     // This should be made more generic around x86 and then support supplying
                     // different kernels instead of treating this as a Linux emulator specifically
                     Some(x) if x == CallOtherOps::SysCall as u64 => {
-                        return self
-                            .linux
-                            .syscall(self.sleigh.as_ref(), memory)
-                            .map_err(|err| Error::DependencyError(Box::new(err)));
+                        return self.kernel.syscall(self.sleigh.as_ref(), memory)
                     }
                     Some(x) if x == CallOtherOps::Lock as u64 => {
                         // Lock instruction. Multithreading not supported so just ignore
