+++ crates/pcode/src/kernel.rs
@@ -1,1 +1,2 @@
 pub mod linux;
+pub mod model;
